Work as intended
The trick here is that I switched from using "sorted_array[0..midpoint-1]"
to "sorted_array[0..midpoint]". The -1 was returning the whole array when
the midpoint was calculated as 0 since 0 - 1 = -1, which for Ruby tells us
to take up to and including the final element in the array.

I don't like the extra check at the end but I couldn't figure out a nice
way to track recursive function calls for items not in the array without
changing the method signature. I'll work on one of those next