Use object instances to track attributes
Ruby is the epitome of OO languages so I should leverage that. The
easiest way of doing this is to use instance vairables to track
properties that transcend the recursive depth. I updated the method
signature too as it doesn't make sense to pass them in when they
are attributes of the instance.

Tricks that I learned:
- I am out of practice with scoping in Ruby and was caught off guard
by the requirement of using 'self' in iterations and if/ternary clauses
- Refactor one thing at a time, even if this means copying and pasting.
When changing the method signature for both classes I copied the
examples into each describe block, refactored one class at a time. It
was easier to keep scope creep out that way.

Things I haven't learned:
- I tried to use 'let(:target) { x }' in the 'each' iterations but this
was always updating target to the final element of the array being
iterated over for all of the examples in the loop. I don't know why this
happened, despite heavy investigation into 'let' vs 'let!'. I don't like
my solution of updating the instance's target value in the expectation
but it works. I also probably wouldn't do so many redundant/fixed tests
if I were to start over