Create general structure for weather data processing
The goal is to parse some data and find the day with the lowest spread.
My current plan is to write a couple of file handlers to process the
data and then pass that in to the weather class. I am thinking I could
use some metaprogramming to create accessor methods for the columns based
of their column titles. Time will tell if this is a good idea or not