Make it run like a café
I built the objects but I never actually implemented them. I had 15
minutes to try this at the end of the day so I pushed to do it in the
easiest way possible. I had some tricks with the initialization values
where I set them to values that would work as opposed to making the code
properly robust